Support raw query against mutable segments
